Added complete code for example
Added code for plotting and displaying policy as well as q_table.
One issue, however, is that it is not working properly. Agent likes
to go left in state 0 which gives reward -10 rather than going
right which gives reward +1. This is a mystery to me. Other than that,
the rest of the policy is as expected.